[Dataflow] Remove concept of 'state_based' reference tables (Recidiviz/recidiviz-data#28790)
## Description of the change

As of Recidiviz/recidiviz-data#28775, there are no longer any Dataflow pipelines that read in
"state-based" reference tables, i.e. tables that do not have
person-level data and must be passed in to each root entity node as side
input.

As a general smell test, it seems good that we're moving in the
direction of removing support for this type of pipeline input. We will
save money by not writing out the same value for every single row from
the metric pipelines when we could just easily join with the reference
view after to get the info we need.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#22528

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3e5ed8020aaaf259c2466a016dd17bd009fed0d0