[US_ND] Tweak StateStaffRolePeriod view (Recidiviz/recidiviz-data#28890)
## Description of the change

There was a bug in the `docstars_staff_role_periods` view in ND that was
causing officer role periods to close prematurely. There were two things
causing this:
1. Data from the `docstars_officers` table was not filtered to only
include rows where an officer was active, so it included all officers in
all updates, which threw off the `last_appearance_datetime` and
`edge_date` fields carried forward from the `staff_from_docstars` CTE.
2. The end_date logic in the `all_periods` CTE was checking that there
was no subsequent critical date, and that the `edge_date <
last_file_update_datetime` to determine that we should close a period.
This means that it was checking if the update to a person's employment
in the row in question happened earlier than the last time we received a
file, and using that information to determine if a person stopped
working. I think the correct logic here is to check if the update in the
row in question happened earlier than the latest time _this person_
appeared in the data, and rely more heavily on the fact that there is no
subsequent critical date. This will hold open periods where the
edge_date is the same as the most recent update about this person, even
when the person does not have a subsequent critical date, which was not
the case before this change.

Tests:

- [x] view query is deterministic
- [x] external IDs are unique
- [x] run_mappings script passes
- [x] test ingest
[pipeline](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-04-12_07_48_58-13169969324237089203;graphView=0?project=recidiviz-123&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ff0e9964a47f5167eb164977efcc0bbb8d8eeaac