Reset thresholds for outliers_staff_count_percent_change_intramonth to default 2 pct (Recidiviz/recidiviz-data#28905)
## Description of the change

The `outliers_staff_count_percent_change_intramonth` validation was
originally configured so that the default threshold was 0% because we
didn't expect any intramonth changes. However, once the validation was
implemented, we discovered some instances in which some low level of
intramonth changes are expected (see Recidiviz/recidiviz-data#28531). Therefore, this PR resets
the thresholds for this validation to be the default 2%.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f58941b0cf6c481cccf62e260e18b94f2f98f805