[US_TN] Closing open out of state supervision periods  (Recidiviz/recidiviz-data#28849)
## Description of the change

Making changes to resolve the known issue in TN where out of state
supervision periods remain open because OffenderMovement is not used in
SPs.

Logic here is taking the last movement from OffenderMovement and if the
MovementType ends in 'OJ' (other jurisdiction) and MovementReason is
'ISC' (interstate compact) we want to close out open supervision
periods. We take the date of that movement to close the period, but if
that date comes before the start date of the period (because TN
reporting has a lag) we make the end date of the period the same as the
start. I added a TODO to revisit the fix of these periods one
OffenderMovement is more heavily incorporated in this view.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#28013

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 154f32bf27ef914c3a7030b317f05f6afb1acd0b