[US_CA] Ingest staff location period and Staff Role Period (Recidiviz/recidiviz-data#28928)
Passes stability checks and passed [an ingest
pipeline](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-04-14_13_02_17-8422241606049276813;graphView=0?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))!

## Description of the change

Query to create staff location periods and staff role periods. The final
periods are
not collapsed -- IE if someones location changes, their role period will
also split, and
even though the role has stayed the same, we will end up with two
consecutive role
periods.

We try to follow logic similar to what was done in
staff_supervisor_periods, but
fragments aren't shared between the two queries yet, so there may be
some slight
differences. TODO(Recidiviz/recidiviz-data#28926)

This means that a PA I (or PA II if they have a caseload) is considered
as working in a
unit if they have a caseload in that unit. If they have a caseload in
multiple units at
the same time (supposedly impossible, but the case for about 56 people
on certain
weeks), we consider that PA I to operate in whichever unit they have the
most clients in
(see Recidiviz/recidiviz-data#28597). I chose not to care about what AgentParole says for the
supervisee
information -- this is because if AgentParole said a PA I was in unit 2,
but they're
supervising 50 clients in unit 3 and 0 in unit 2, they're clearly in
unit 3.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#28212, Recidiviz/recidiviz-data#24604

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7af5abac0f7e6bfbde7b8f5712bda604fb0af714