[Insights] Add logic to identify top x% of officers to product view and querier (Recidiviz/recidiviz-data#28494)
## Description of the change

This PR updates product views and the querier method in order to
identify if an officer should be highlighted as being in the top x% of a
metric. This does so by adding logic in the `metric_benchmarks` product
view to calculate the percentile value to compare an individual rate to,
then adding those fields and an `is_top_x_pct` field to the
`supervision_officer_outlier_status` view. Finally, the result is
communicated to the client by adding a `top_x_pct_metrics` field to the
`SupervisionOfficerEntity`, which is a list of objects that contains the
metric id and the `x` value for `top x%` for all metrics that an officer
should be "highlighted" for.

I tested this out by adding highlight_percentiles to the metric configs
for an adverse and favorable metric in IX, then loaded these views into
`alexa_28252_`. I'll remove these in the config before merging. I feel
somewhat confident that this works because if you run this query,
`should_highlight=True` only when an officer's status is 'MET' which is
correct behavior.

Note: I felt it would be redundant to add the new fields on the
`MetricBenchmark` schema entity to the results of the /benchmarks
endpoint, but I can do so if it makes sense to keep these values there
and not in the new field on `SupervisionOfficerEntity`. We should also
create a follow up to add this field to the CA config once implemented
(Recidiviz/recidiviz-data#28710) and a follow up to add the new fields to the
`_outlier_status_archive` view (Recidiviz/recidiviz-data#28711)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#28252

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 977cb803e728491ddc555f0e61a2629e5a589cd9