[US_CA] Include ParoleUnit when computing changes in supervision periods (Recidiviz/recidiviz-data#28630)
## Description of the change

[Passes
pipelines](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-04-15_11_42_54-4413793700132980098;graphView=1?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))!

We were seeing an issue where clients were not being moved if their unit
changed within the same district. This was because we accidentally
didn't include the ParoleUnit in the concatenation of fields which
allows to detect changes in location. This PR adds ParoleUnit so if the
unit changes, the concatenated field changes.

I changed a test to cover this case, but didn't need to change the
result since in both the old test and the new test, the result is that
the change in units is correctly captured. However, if you take the
change to this fixture data and try running the test in main, you'll see
that it fails.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#28542

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7157882c21cdd911150a7850e12c56ddbe3201f0