fix syntax error in metric_benchmarks (Recidiviz/recidiviz-data#28974)
## Description of the change

When https://github.com/Recidiviz/recidiviz-data/pull/28494 was being
developed, a config was added with `top_x_pct` to show that it worked,
but when that config was removed, it introduced a syntax error because
the cte was fully empty. This adds null placeholders for now so that it
compiles but doesn't join to anything.

Loaded just this view into sandbox at `dana_outliers`, verified that it
has the same # of rows as the current `metric_benchmarks` view.

This unblocks today's alpha deploy.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f5ef08087087f942e4ad352cb33eebf97fcec845