Add Firestore auth variable + instructions on how to use it (Recidiviz/recidiviz-data#28973)
## Description of the change
- Added a `firestoreAuthorized` MobX observable variable to the
`RootStore` that is set to true once firestore authentication has
completed successfully, allowing other parts of the application to
subscribe to this change. The change was taken from Ian's prototype
branch.
- Without this change, if other parts of the application attempt to
connect to firestore, there is no way for them to check whether
authentication has been successful, which may result in a race
condition.
- Added a small bit to the README about how to use this new feature.

## Type of change
Feature

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#27582

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: bfa72c31529e60d204530c0ae324243130cfebc4