Consider supervision period end reasons in sessions inference when overlapping with incarceration end reasons (Recidiviz/recidiviz-data#27882)
## Description of the change
This PR adds an additional join to
`compartment_sub_sessions_preprocessed` to pull in an additional end
reason in cases where there is a supervision termination that occurs on
the same day as an incarceration release. This supervision reason was
previously being dropped entirely because it doesn't join to sessions
post deduplication.

Logic is then implemented to recategorize the end reason field in cases
where there is a supervision end reason that indicates a revocation or
admission to prison that overlaps with an incarceration end reason that
indicates a temporary custody release.

Downstream this results in us no longer inferring LIBERTY in cases where
there is a gap following that end reason. Those inferred periods would
instead be labeled INTERNAL_UNKNOWN. Further downstream, this means that
we will no longer be incorrectly inferring a completion date in
sentencing data, which is the main motivation for this update.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#27022

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e11a141e89a2322126e425bec95eba4c42daa498