Use email to identify Workflows users in Workflows Usage dashboard (Recidiviz/recidiviz-data#28722)
## Description of the change

This PR makes the relevant changes to identify Workflows users via email
instead of external id in the Workflows Usage dashboard.

This happens by adding the `email` column to
`reidentified_dashboard_users`, then outputting the email in
`workflows_officer_events` and `workflows_usage`. The latter view is the
source table for the existing usage dashboard in Looker. Looker pulls
the staff names from `reference_views.state_staff_with_names`, so I also
add the email to the output of this view so that we can join on the
email in this view and `workflows_usage`. *This requires a follow up in
Looker to update the join. However, one alternate option here is to
output the clean name in `workflows_usage` and eliminate the need for
the reference view in Looker.*

I loaded the new views into sandbox with prefix `alexa_5211`. Quick
notes on the impact of the change: There are 3k total events that
previously had no external_id, but now have an associated email. These,
previously unidentified, events happened for ~30 distinct users.

```
SELECT
  state_code,
  -- event_type,
  COUNT(DISTINCT email)
FROM `recidiviz-staging.alexa_5211_analyst_data.workflows_officer_events`
WHERE
  officer_id IS NULL
  AND email IS NOT NULL
GROUP BY 1
ORDER BY 1
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d460441ac6e6e1496040e2c430d3a71160a33322