[BigQuery] Cluster most recent dataflow metrics by state code (Recidiviz/recidiviz-data#29034)
## Description of the change

We got the following recommendation in the BQ UI. It looks like we
currently scan 800 TiB of data per month, so 100TiB is a good amount of
savings. We might as well turn it on for all most recent dataflow
metrics.

![Screenshot 2024-04-16 at 1 34
44 PM](https://github.com/Recidiviz/recidiviz-data/assets/3762913/b8d3f6de-28f7-4848-ab00-6b91cc461c27)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

N/A

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 37662221f22df2b65445be3ab86422354705e9ea